feat: enterpriseStart command to create or start a simulation
Motivation:
We wan't a single command to create / start simulation, interactively or not.

Modifications:
If simulationId is configured, upload to simulation package and start the simulation.
If not, and batch mode option is configured, try to create with default values a package and a simulation using it, then upload to package and start simulation
Otherwise, start interactive creation or start of a simulation

Ref: MISC-318