chore: Fix warnings about task evaluations inside anonymous functions
Motivation:
The SBT task linter warns when a task is evaluated in an anonymous function, because the task will be evaluated regardless of whether the function is actually executed, which is counter-intuitive.

Modifications:
Move such task evaluations to a val before the function call. This makes it clear when it actually gets evaluated.

Result:
No warnings in the build.